// src/app/models/Permissions.ts
export class Permissions {
  add: boolean = false;
  add_url: string = null;
  view: boolean = false;
  view_url: string = null;
  inside_view: boolean = false;
  inside_view_url: string = null;
  edit: boolean = false;
  edit_url: string = null;
  edit_current_form: boolean = false;
  deactivate: boolean = false;
  deactivate_url: string = null;
  activate: boolean = false;
  activate_url: string = null;
  delete: boolean = false;
  delete_url: string = null;
  changeStatusDelete: boolean = false;
  changeStatusDelete_url: string = null;
  forwardDestinationDelete: boolean = false;
  forwardDestinationDelete_url: string = null;

  stationUserDelete: boolean = false;
  stationUserDelete_url: string = null;

  changeStatusForm: boolean = false;
  asssignToOfficer: boolean = false;
  changeStatusList: boolean = false;
  forwardToDestination: boolean = false;
  forwardToDestinationList: boolean = false;
  sendSMSToUser: boolean = false;
  sendSMSToUserList: boolean = false;

  stationUser: boolean = false;
  stationUserList: boolean = false;

  sectionUser: boolean = false;
  sectionUserList: boolean = false;

  dspUser: boolean = false;
  dspUserList: boolean = false;

  sdpoUser: boolean = false;
  sdpoUserList: boolean = false;

  subdivisionUser: boolean = false;
  subdivisionUserList: boolean = false;

  circleUser: boolean = false;
  circleUserList: boolean = false;

  courtDetailsForm: boolean = false;
  courtDetailsList: boolean = false;

  grSection: boolean = false;
  grSectionList: boolean = false;
  grSectionDelete: boolean = false;
  grSectionDelete_url: string = null;

  grSectionDetailsForm: boolean = false;
  grSectionDetailsList: boolean = false;

  procecution: boolean = false;
  procecutionList: boolean = false;
  procecutionDelete: boolean = false;
  procecutionDelete_url: string = null;

  procecutionDetailsForm: boolean = false;
  procecutionDetailsList: boolean = false;

  supplierProducts: boolean = false;
  supplierProductsList: boolean = false;

  addForm: boolean = false;
  editForm: boolean = false;
  viewForm: boolean = false;
  addFormUrl: string = null;
  editFormUrl: string = null;
  viewFormUrl: string = null;

  download_cert: boolean = false;
  uploadCertPermission: boolean = false;
  activeInactive: boolean = false;
  fetch:boolean = false;
  fetch_url: string = null;

  fetch1:boolean = false;
  fetch_url1: string = null;

  fetch2:boolean = false;
  fetch_url2: string = null;

}